finishes search refactor
